requirements.txt
python-decouple
jinja2

